Relax httpx dependency from ^0.23.0 to >=0.23.0
The carret prevents any newer 'minor' releases of httpx to be used,
because the 'major' version is 0. This makes it difficult to upgrade
other dependencies of ScopeSim like notebook, because those depend on
newer versions of httpx.

It is probably fine to just use >= for httpx, as we only use basic
features that will probably not be deprecated or break.